pyproject.toml
[project]
name = "mts"
version = "3.0.0"
description = "MTS - Autonomous AI Trading Agent for Hyperliquid"
authors = []
requires-python = ">=3.9,<3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "pydantic>=2.6.3",
    "pydantic-settings>=2.2.1",
    "pydantic-ai>=0.0.7",
    "anthropic>=0.18.1",
    "python-dotenv>=1.0.1",
    "eth-account>=0.11.0",
    "loguru>=0.7.2",
    "rich>=13.7.0",
    "pandas>=2.2.1",
    "numpy>=1.26.4",
    "plotly>=5.19.0",
    "httpx>=0.27.0",
    "hyperliquid",
    "google-adk",
]

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.2"
pytest-asyncio = ">=0.23.5"
pytest-cov = ">=4.1.0"
black = ">=24.2.0"
ruff = ">=0.3.2"
mypy = ">=1.8.0"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true